refactor: reorganize src/ into module directories
Move flat source files into a clean module hierarchy following
the compiler pipeline (source → tokens → AST → analysis → codegen):

src/
  lexer/      token.rs, lexer.rs (was flat in src/)
  ast/        split into expr.rs, stmt.rs, decl.rs
  parser/     mod.rs (was parser.rs)
  analysis/   checker.rs, environment.rs, types.rs
  codegen/    compiler.rs, transpiler/sql.rs
  error.rs    lib.rs   main.rs   runtime.c

Closes #8
