+++ src/analysis/mod.rs
@@ -1,0 +1,7 @@
+pub mod checker;
+pub mod environment;
+pub mod types;
+
+pub use checker::TypeChecker;
+pub use environment::Environment;
+pub use types::Type;
