+++ src/ast/mod.rs
@@ -1,0 +1,25 @@
+pub mod expr;
+pub mod stmt;
+pub mod decl;
+
+pub use expr::*;
+pub use stmt::*;
+pub use decl::*;
+
+use crate::lexer::token::Token;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct Span {
+    pub line: usize,
+    pub col: usize,
+}
+
+impl Span {
+    pub const fn zero() -> Self {
+        Self { line: 0, col: 0 }
+    }
+
+    pub fn from_token(tok: &Token) -> Self {
+        Self { line: tok.line, col: tok.col }
+    }
+}
