+++ src/lexer/mod.rs
@@ -1,0 +1,5 @@
+pub mod token;
+pub mod lexer;
+
+pub use token::{Token, TokenKind};
+pub use lexer::Lexer;
